storieの修正
